update bot code to use persisted data and update persistence if necessary refactoring
